requirements.txt
imageio==2.9.0
matplotlib==3.3.2
opencv-python==4.4.0.44
pandas==1.1.3
Pillow==7.2.0
progressbar==2.5
scikit-image==0.16.2
scikit-learn==0.20.2
scipy==1.1.0
timm==0.4.5